+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## unreleased version -- v0.10.7
+### Added
+- Docker Compose file to set up local dev environment
+- Updated README steps for setting up local dev environment 
+### Fixed
+- Updated unit tests for developmentSetup process
+
 ## 2022-09-07 -- v0.10.6
 ### Added
 - Government document search filter